Revert "Move Nativebase mappings out of folder"
This reverts commit 52c9e4867ae8597e9a2303a3d63a7094f6ecc8b9.
